Update dice.py
https://www.youtube.com/watch?v=5k-gZQiHdAA
The following code counts the number of dots on both white and colored die through the use of a camera operated by a Raspberry Pi. The camera takes time to adjust to the lighting for seeing colored die due to the HSV filtering that's being processed through the Pi. Upon adjustment, the system can accurately count the number of dots.